CODE RUB: Integrate KeyVault and SecureData services in Startup
This commit includes the addition of `LHDS.Core.Brokers.KeyVaults` and `LHDS.Core.Services.Foundations.SecureDatas` namespaces to the `Startup.cs` file, enabling the use of classes and methods within these namespaces. The `IKeyVaultSecretBroker` and `ISecureDataService` services were also added to the dependency injection container, replacing the previously commented out code. In the `KeyVaultSecretBroker.cs` file, the constructor was modified to accept an `IConfiguration` object instead of a string for the KeyVaultUri, allowing the KeyVaultUri to be retrieved from the application's configuration settings. A new `Configuration` property was added to the `KeyVaultSecretBroker` class for retrieving the KeyVaultUri. No changes were made to the `CreateOrUpdateKeyVaultSecretAsync` method.