Updated Startup.cs and subscriberAgreementTable.tsx, added SubscriberAgreementsController.cs
Updated the `Startup.cs` file to include the `SubscriberAgreements` namespace and added `SubscriberAgreements` to the Entity Framework builder. The `subscriberAgreementTable.tsx` file has been updated to import `ButtonBase`, `FontAwesomeIcon`, and `faPlusCircle`, and a new `ButtonBase` component has been added to the `SubscriberAgreementTable` component. A new `SubscriberAgreementsController.cs` file has been created with a `SubscriberAgreementsController` class that includes methods for creating, retrieving, updating, and deleting `SubscriberAgreement` objects, using the `SubscriberAgreementService` to interact with the database.